# src/main/resources/openapi.yml
openapi: '3.0.1'
info:
  version: 1.0.0
  title: Money Transfer Service API
servers:
  - url: /v1
    description: localhost
components:
  schemas:
    JsonProblem:
      properties:
        type:
          type: string
          format: uri
          default: 'about:blank'
        status:
          type: integer
        title:
          type: string
        detail:
          type: string
        instance:
          type: string
          format: uri
    JsonProblem.Invalid:
      allOf:
        - $ref: '#/components/schemas/JsonProblem'
        - properties:
            violations:
              type: object
              additionalProperties:
                type: array
                items:
                  type: string
...
    PageMetadata:
      properties:
        total:
          type: integer
        size:
          type: integer
        pages:
          type: integer
        page:
          type: integer
...
    CurrencyId:
      type: string
      minLength: 2
      maxLength: 10
    Rate:
      properties:
        source:
          $ref: '#/components/schemas/CurrencyId'
        target:
          $ref: '#/components/schemas/CurrencyId'
        value:
          type: string
          format: decimal
    RatePage:
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/Rate'
        metadata:
          $ref: '#/components/schemas/PageMetadata'
...
    DecimalValueContainer:
      required: [value]
      properties:
        value:
          type: string
          format: decimal
...
  parameters:
...
    query.page:
      in: query
      name: page
      schema:
        type: integer
        default: 1
    query.size:
      in: query
      name: size
      schema:
        type: integer
        default: 32
  requestBodies:
    DecimalValue:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/DecimalValueContainer'
...
  responses:
    404:
      description: Requested resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/JsonProblem'
    400.Invalid:
      description: Invalid request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/JsonProblem.Invalid'
...
    Rate:
      description: Single exchange rate
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Rate'
    RatePage:
      description: Exchange rate listing page
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/RatePage'
...
paths:
...
  /rate:
    get:
      summary: Rate listing
      parameters:
        - $ref: '#/components/parameters/query.page'
        - $ref: '#/components/parameters/query.size'
      responses:
        200:
          $ref: '#/components/responses/RatePage'
        404:
          $ref: '#/components/responses/404'
  /rate/{source}/{target}:
    parameters:
      - name: source
        in: path
        required: true
        schema:
          type: string
      - name: target
        in: path
        required: true
        schema:
          type: string
    get:
      summary: Exchange rate fetching
      responses:
        200:
          $ref: '#/components/responses/Rate'
        404:
          $ref: '#/components/responses/404'
    put:
      summary: Exchange rate setting
      requestBody:
        $ref: '#/components/requestBodies/DecimalValue'
      responses:
        200:
          $ref: '#/components/responses/Rate'
        400:
          $ref: '#/components/responses/400.Invalid'
...